puliendo bullets

+++ Assets/Scripts/Player/Bullet/LinealBullet.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class LinealBullet : MonoBehaviour{
-    public int speed = 2;
+    public float speed = 2;
     public Vector3 direction = Vector3.right;
 
     public int damage = 10; //valor de daño de esta bala
